template error

+++ resources/views/pdf/send_email_template.blade.php
@@ -671,9 +671,9 @@
                             <td style="padding: 0 0 5px; margin: 0; border-bottom: 1px solid #999;">
                                 <p style="marging: 0; padding: 0">{{ $product->quantity }}</p>
                             </td>
-                            @if($request->format != 'without_values') 
+                            @if((float)$request->format != 'without_values') 
                                 <td style="padding: 0 0 5px; margin: 0; border-bottom: 1px solid #999;">
-                                    <p style="marging: 0; padding: 0">{{ (int)$product->discount }}</p>
+                                    <p style="marging: 0; padding: 0">{{ (float)($product->discount }}</p>
                                 </td>
                                 <td style="padding: 0 0 5px; margin: 0; border-bottom: 1px solid #999;">
                                     <p style="marging: 0; padding: 0">{{ $product->base_price }}</p>
